fix(*): 修复数据权限查询bug

+++ sys_callback.go
@@ -7,12 +7,12 @@ import (
 
 // CreateCallback
 var CreateCallback = func(scope *gorm.Scope) {
-	if v, ok := GetValue(PrisDescKey); ok {
+	if v, ok := GetValue(PrisDescKey); ok && v != nil {
 		desc := v.(*PrivilegesDesc)
-		if desc != nil && !scope.HasError() {
+		if desc.IsValid() && !scope.HasError() {
 			if orgIDField, ok := scope.FieldByName("OrgID"); ok {
 				if orgIDField.IsBlank {
-					orgIDField.Set(desc.OrgID)
+					orgIDField.Set(desc.SignOrgID)
 				}
 			}
 			if createdByField, ok := scope.FieldByName("CreatedByID"); ok {
@@ -28,9 +28,9 @@ var CreateCallback = func(scope *gorm.Scope) {
 
 // DeleteCallback
 var DeleteCallback = func(scope *gorm.Scope) {
-	if v, ok := GetValue(PrisDescKey); ok {
+	if v, ok := GetValue(PrisDescKey); ok && v != nil {
 		desc := v.(*PrivilegesDesc)
-		if desc != nil && !scope.HasError() {
+		if desc.IsValid() && !scope.HasError() {
 			var extraOption string
 			if str, ok := scope.Get("gorm:delete_option"); ok {
 				extraOption = fmt.Sprint(str)
@@ -59,9 +59,9 @@ func addExtraSpaceIfExist(str string) string {
 
 // UpdateCallback
 var UpdateCallback = func(scope *gorm.Scope) {
-	if v, ok := GetValue(PrisDescKey); ok {
+	if v, ok := GetValue(PrisDescKey); ok && v != nil {
 		desc := v.(*PrivilegesDesc)
-		if desc != nil {
+		if desc.IsValid() {
 			scope.SetColumn("UpdatedByID", desc.UID)
 		}
 	}
@@ -69,10 +69,24 @@ var UpdateCallback = func(scope *gorm.Scope) {
 
 // QueryCallback
 var QueryCallback = func(scope *gorm.Scope) {
-	if v, ok := GetValue(PrisDescKey); ok {
-		desc := v.(*PrivilegesDesc)
-		if desc != nil {
-			scope.Search.Where("(org_id IS NULL) OR (org_id in (?)) OR (created_by_id = ?)", desc.ReadableOrgIDs, desc.UID)
+	rawDesc, _ := GetValue(PrisDescKey)
+	rawValues, _ := GetValue(ValuesKey)
+
+	if !IsBlank(rawDesc) {
+		if !IsBlank(rawValues) {
+			values := make(Values)
+			values = *(rawValues.(*Values))
+			if _, ok := values[IgnoreAuthKey]; ok {
+				return
+			}
+		}
+		desc := rawDesc.(*PrivilegesDesc)
+		if desc.NotRootUser() {
+			_, hasOrgIDField := scope.FieldByName("OrgID")
+			_, hasCreatedByIDField := scope.FieldByName("CreatedByID")
+			if hasOrgIDField && hasCreatedByIDField {
+				scope.Search.Where("(org_id IS NULL) OR (org_id in (?)) OR (created_by_id = ?)", desc.ReadableOrgIDs, desc.UID)
+			}
 		}
 	}
 }
